+++ src/Sites/Playground/Up.php
@@ -1,31 +1,21 @@
 <?php
 
-namespace Portal\Sites\Portal;
+namespace Portal\Sites\Playground;
 
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
-use OriginEngine\Contracts\Helpers\Settings\SettingRepository;
 use OriginEngine\Helpers\Directory\Directory;
 use OriginEngine\Helpers\Env\EnvRepository;
-use OriginEngine\Helpers\IO\IO;
-use OriginEngine\Helpers\Terminal\Executor;
 use OriginEngine\Pipeline\PipelineConfig;
 use OriginEngine\Pipeline\PipelineHistory;
-use OriginEngine\Pipeline\Tasks\Files\OverwriteFile;
-use OriginEngine\Pipeline\Tasks\Git\CloneGitRepository;
 use OriginEngine\Pipeline\Pipeline;
-use OriginEngine\Pipeline\Tasks\Files\CopyFile;
 use OriginEngine\Pipeline\Tasks\EditEnvironmentFile;
-use OriginEngine\Pipeline\Tasks\LaravelSail\GenerateApplicationKey;
 use OriginEngine\Pipeline\Tasks\LaravelSail\InstallNpmDependencies;
 use OriginEngine\Pipeline\Tasks\LaravelSail\MigrateDatabase;
 use OriginEngine\Pipeline\Tasks\LaravelSail\RunNpmScript;
-use OriginEngine\Pipeline\Tasks\LaravelSail\SeedLaravel;
-use OriginEngine\Pipeline\Tasks\Utils\Closure;
 use OriginEngine\Pipeline\Tasks\WaitForDocker;
-use Portal\Tasks\GeneratePassportApiKeys;
 
-class Install extends Pipeline
+class Up extends Pipeline
 {
 
     public function __construct()
@@ -52,9 +42,8 @@ class Install extends Pipeline
     public function tasks(): array
     {
         return [
-            'clone' => (new CloneGitRepository('git@github.com:bristol-su/portal', 'develop')),
             'composer-install' => new \OriginEngine\Pipeline\Tasks\LaravelSail\InstallComposerDependencies('80'),
-            'create-local-env-file' => new CopyFile('.env.sail.example', '.env'),
+
             'check-ports-free' => new \OriginEngine\Pipeline\Tasks\CheckPortsAreFree(
                 '.env',
                 [
@@ -62,7 +51,7 @@ class Install extends Pipeline
                     'FORWARD_DB_PORT' => 'database',
                     'FORWARD_MAILHOG_PORT' => 'mail',
                     'FORWARD_MAILHOG_DASHBOARD_PORT' => 'mail dashboard',
-                    'FORWARD_REDIS_PORT' => 'redis',
+                    'FORWARD_REDIS_PORT' => 'redis'
                 ],
                 false),
 
@@ -70,23 +59,13 @@ class Install extends Pipeline
 
             'bring-sail-up' => new \OriginEngine\Pipeline\Tasks\LaravelSail\BringSailEnvironmentUp(),
 
-            'wait-for-docker' => new WaitForDocker(),
+            'wait-for-docker' => (new WaitForDocker())->setUpName('Waiting for Docker. This may take a minute.'),
 
-            'create-application-key' => new GenerateApplicationKey('local', '.env'),
-
-            'migrate-db' => new MigrateDatabase('local'),
-
-            'create-npm-registration' => new OverwriteFile(
-                '.npmrc',
-                sprintf('//npm.pkg.github.com/:_authToken=%s', app(SettingRepository::class)->get('github-npm-token'))
-                . PHP_EOL . '@bristol-su:registry=https://npm.pkg.github.com'
-            ),
+            'migrate-local-db' => new MigrateDatabase('local'),
 
             'install-npm-dependencies' => new InstallNpmDependencies(),
 
             'compile-assets' => new RunNpmScript('dev'),
-
-            'generate-api-keys' => new GeneratePassportApiKeys()
 
         ];
     }